feat: use the same reducer on mobile and on browser
